+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased](https://github.com/schemathesis/schemathesis/compare/v4.0.12...HEAD) - TBD
 
+### :bug: Fixed
+
+- Honor filters defined via a config file in `pytest` integration.
+
 ## [4.0.12](https://github.com/schemathesis/schemathesis/compare/v4.0.11...v4.0.12) - 2025-07-21
 
 ### :wrench: Changed